Viewport values should be integers

+++ gtk_ui/src/canvas/viewport.rs
@@ -4,20 +4,20 @@
 #[derive(Copy, Clone, PartialEq, Debug)]
 pub struct CanvasViewport {
     /// The total width of the canvas
-    pub width: f32,
+    pub width: i32,
 
     /// The total height of the canvas
-    pub height: f32,
+    pub height: i32,
 
     /// The viewport origin X coordinate
-    pub viewport_x: f32,
+    pub viewport_x: i32,
 
     /// The viewport origin Y coordinate
-    pub viewport_y: f32,
+    pub viewport_y: i32,
 
     /// The viewport width
-    pub viewport_width: f32,
+    pub viewport_width: i32,
 
     /// The viewport height
-    pub viewport_height: f32
+    pub viewport_height: i32
 }
